src/main/java/juliancambraia/springframework/converters/CategoryToCategoryCommand.java
package juliancambraia.springframework.converters;

import juliancambraia.springframework.commands.CategoryCommand;
import juliancambraia.springframework.domain.Category;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

/**
 * @author juliancambraia
 * @implNote Cria classe de conversão (Mapper) entre a entidade Category para CategoryCommand
 */

@Component
public class CategoryToCategoryCommand implements Converter<Category, CategoryCommand> {

    @Synchronized
    @Nullable
    @Override
    public CategoryCommand convert(Category category) {
        if (category == null) {
            return null;
        }
        final CategoryCommand categoryCommand = new CategoryCommand();
        categoryCommand.setId(category.getId());
        categoryCommand.setDescription(category.getDescription());

        return categoryCommand;
    }
}
